GER-539: add getLfsCachedInfoForReposiory
Added REST API ability to request is an oid cached for a given repo, not just rep group.
Added information to the returned LfsPush calls, so that if an item is cached, or is being added, and we just use the projectName that we return the relevant projectId, and replication group Id/Name to be able to cache based on rep group moving forward.