GER-539: Updated LfsReplicationRequestBuilder
The LfsReplicationRequestBuilder, has now had its implementation deleted in favour of a generic implmentation we have in the gerrit-gitms-shared project which can deal with any HttpRequest to a REST Api and has general handling for objects on success and failure.  This allows for us to centrate on the logic of the request and keep a central class for the construction and issuing of the request.